Move bboxFromVertices and friends to polygon.c
