+++ src/apps/testapps/testPolygon.c
@@ -15,6 +15,7 @@
  */
 
 #include <stdlib.h>
+#include "bbox.h"
 #include "constants.h"
 #include "geoCoord.h"
 #include "h3Index.h"
@@ -93,6 +94,79 @@ TEST(loopContainsPointTransmeridian) {
     t_assert(loopContainsPoint(&transMeridianGeofence, &bbox,
                                &eastPointOutside) == false,
              "does not contain outside point to the east of the antimeridian");
+}
+
+TEST(noVertices) {
+    const BBox expected = {0.0, 0.0, 0.0, 0.0};
+
+    BBox result;
+    bboxFromVertices(NULL, 0, &result);
+
+    t_assert(bboxEquals(&result, &expected), "Got expected bbox");
+}
+
+TEST(bboxFromGeofence) {
+    GeoCoord verts[] = {{0.8, 0.3}, {0.7, 0.6}, {1.1, 0.7}, {1.0, 0.2}};
+
+    Geofence geofence;
+    geofence.verts = verts;
+    geofence.numVerts = 4;
+
+    const BBox expected = {1.1, 0.7, 0.7, 0.2};
+
+    BBox result;
+    bboxFromGeofence(&geofence, &result);
+    t_assert(bboxEquals(&result, &expected), "Got expected bbox");
+}
+
+TEST(bboxesFromGeoPolygon) {
+    GeoCoord verts[] = {{0.8, 0.3}, {0.7, 0.6}, {1.1, 0.7}, {1.0, 0.2}};
+
+    Geofence geofence;
+    geofence.verts = verts;
+    geofence.numVerts = 4;
+
+    GeoPolygon polygon;
+    polygon.geofence = geofence;
+    polygon.numHoles = 0;
+
+    const BBox expected = {1.1, 0.7, 0.7, 0.2};
+
+    BBox* result = calloc(sizeof(BBox), 1);
+    bboxesFromGeoPolygon(&polygon, result);
+    t_assert(bboxEquals(&result[0], &expected), "Got expected bbox");
+
+    free(result);
+}
+
+TEST(bboxesFromGeoPolygonHole) {
+    GeoCoord verts[] = {{0.8, 0.3}, {0.7, 0.6}, {1.1, 0.7}, {1.0, 0.2}};
+
+    Geofence geofence;
+    geofence.verts = verts;
+    geofence.numVerts = 4;
+
+    // not a real hole, but doesn't matter for the test
+    GeoCoord holeVerts[] = {{0.9, 0.3}, {0.9, 0.5}, {1.0, 0.7}, {0.9, 0.3}};
+
+    Geofence holeGeofence;
+    holeGeofence.verts = holeVerts;
+    holeGeofence.numVerts = 4;
+
+    GeoPolygon polygon;
+    polygon.geofence = geofence;
+    polygon.numHoles = 1;
+    polygon.holes = &holeGeofence;
+
+    const BBox expected = {1.1, 0.7, 0.7, 0.2};
+    const BBox expectedHole = {1.0, 0.9, 0.7, 0.3};
+
+    BBox* result = calloc(sizeof(BBox), 2);
+    bboxesFromGeoPolygon(&polygon, result);
+    t_assert(bboxEquals(&result[0], &expected), "Got expected bbox");
+    t_assert(bboxEquals(&result[1], &expectedHole), "Got expected hole bbox");
+
+    free(result);
 }
 
 END_TESTS();
